:building_construction: (RUMM-160): Move forgery factory to dedicated package
