+++ dd-sdk-android/src/test/kotlin/com/datadog/android/core/internal/data/file/DeferredWriterTest.kt
@@ -5,7 +5,7 @@ import com.datadog.android.core.internal.data.DataMigrator
 import com.datadog.android.core.internal.data.Writer
 import com.datadog.android.core.internal.threading.AndroidDeferredHandler
 import com.datadog.android.core.internal.threading.LazyHandlerThread
-import com.datadog.android.log.forge.Configurator
+import com.datadog.android.utils.forge.Configurator
 import com.datadog.tools.unit.annotations.TestTargetApi
 import com.datadog.tools.unit.extensions.ApiLevelExtension
 import com.datadog.tools.unit.extensions.SystemOutputExtension
