+++ dd-sdk-android/src/test/kotlin/com/datadog/android/core/internal/data/file/FileExtensionsTest.kt
@@ -1,6 +1,6 @@
 package com.datadog.android.core.internal.data.file
 
-import com.datadog.android.log.forge.Configurator
+import com.datadog.android.utils.forge.Configurator
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.spy
 import com.nhaarman.mockitokotlin2.whenever
