+++ dd-sdk-android/src/test/kotlin/com/datadog/android/core/internal/net/GzipRequestInterceptorTest.kt
@@ -6,7 +6,7 @@
 
 package com.datadog.android.core.internal.net
 
-import com.datadog.android.log.forge.Configurator
+import com.datadog.android.utils.forge.Configurator
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.argumentCaptor
 import com.nhaarman.mockitokotlin2.doReturn
