+++ dd-sdk-android/src/test/kotlin/com/datadog/android/core/internal/net/NetworkTimeInterceptorTest.kt
@@ -7,7 +7,7 @@
 package com.datadog.android.core.internal.net
 
 import com.datadog.android.core.internal.time.MutableTimeProvider
-import com.datadog.android.log.forge.Configurator
+import com.datadog.android.utils.forge.Configurator
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.anyOrNull
 import com.nhaarman.mockitokotlin2.argumentCaptor
