+++ dd-sdk-android/src/test/kotlin/com/datadog/android/core/internal/time/DatadogTimeProviderTest.kt
@@ -8,7 +8,7 @@ package com.datadog.android.core.internal.time
 
 import android.content.Context
 import android.content.SharedPreferences
-import com.datadog.android.log.forge.Configurator
+import com.datadog.android.utils.forge.Configurator
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.argumentCaptor
 import com.nhaarman.mockitokotlin2.doReturn
