+++ dd-sdk-android/src/test/kotlin/com/datadog/android/domain/LogSerializerTest.kt
@@ -3,10 +3,10 @@ package com.datadog.android.domain
 import com.datadog.android.BuildConfig
 import com.datadog.android.Datadog
 import com.datadog.android.log.assertj.JsonObjectAssert.Companion.assertThat
-import com.datadog.android.log.forge.Configurator
 import com.datadog.android.log.internal.domain.Log
 import com.datadog.android.log.internal.domain.LogSerializer
 import com.datadog.android.log.internal.user.UserInfo
+import com.datadog.android.utils.forge.Configurator
 import com.google.gson.JsonObject
 import com.google.gson.JsonParser
 import com.google.gson.JsonPrimitive
