+++ dd-sdk-android/src/test/kotlin/com/datadog/android/log/internal/net/BroadcastReceiverNetworkInfoProviderTest.kt
@@ -14,7 +14,7 @@ import android.net.NetworkInfo as AndroidNetworkInfo
 import android.os.Build
 import android.telephony.TelephonyManager
 import com.datadog.android.log.assertj.NetworkInfoAssert.Companion.assertThat
-import com.datadog.android.log.forge.Configurator
+import com.datadog.android.utils.forge.Configurator
 import com.datadog.tools.unit.annotations.TestTargetApi
 import com.datadog.tools.unit.extensions.ApiLevelExtension
 import com.nhaarman.mockitokotlin2.doReturn
