+++ dd-sdk-android/src/test/kotlin/com/datadog/android/log/internal/net/CallbackNetworkInfoProviderTest.kt
@@ -12,7 +12,7 @@ import android.net.Network
 import android.net.NetworkCapabilities
 import android.os.Build
 import com.datadog.android.log.assertj.NetworkInfoAssert.Companion.assertThat
-import com.datadog.android.log.forge.Configurator
+import com.datadog.android.utils.forge.Configurator
 import com.datadog.tools.unit.annotations.TestTargetApi
 import com.datadog.tools.unit.extensions.ApiLevelExtension
 import com.nhaarman.mockitokotlin2.any
