+++ dd-sdk-android/src/test/kotlin/com/datadog/android/log/internal/system/BroadcastReceiverSystemInfoProviderTest.kt
@@ -11,7 +11,7 @@ import android.content.Intent
 import android.os.BatteryManager
 import android.os.PowerManager
 import com.datadog.android.log.assertj.SystemInfoAssert.Companion.assertThat
-import com.datadog.android.log.forge.Configurator
+import com.datadog.android.utils.forge.Configurator
 import com.datadog.tools.unit.extensions.ApiLevelExtension
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.doAnswer
